Add second function for saving coordinates
Add second function for saveing coordinates (case with fixed number of evaluation points. This is done because if the lenghts of lists is same, them after saving into mat file data is stored as array Int64 instead of cell type.
Link to the issue: https://github.com/scipy/scipy/issues/3756